Created the blueprints folder and moved the routes in respective files
